refactor: move domyland funcs to own module
