+++ src/controller/order_classification/vysota/order_classification_controller.py
@@ -1,14 +1,23 @@
-from fastapi import APIRouter, status, Depends
-from fastapi import Response
+from fastapi import APIRouter, status, Depends, Response
 from fastapi_versioning import version
+from loguru import logger
 from sqlmodel import Session
 
 from controller.user.user_controller import get_current_user
 from infra.database import get_session
-from model.order_classification import order_classification_model, order_classification_config_model, order_get_info
-from scheme.order_classification.order_classification_config_scheme import OrderClassificationConfig
-from scheme.order_classification.order_classification_history_scheme import OrderClassificationRecord
-from scheme.order_classification.order_classification_scheme import OrderClassificationRequest
+from model.order_classification import (
+    order_classification_model,
+    order_classification_config_model,
+)
+from scheme.order_classification.order_classification_config_scheme import (
+    OrderClassificationConfig,
+)
+from scheme.order_classification.order_classification_history_scheme import (
+    OrderClassificationRecord,
+)
+from scheme.order_classification.order_classification_scheme import (
+    OrderClassificationRequest,
+)
 from scheme.user.user_scheme import UserRead
 
 router = APIRouter()
@@ -21,8 +30,6 @@ def get_classification_config_by_id(
     session: Session = Depends(get_session),
     current_user: UserRead = Depends(get_current_user),
 ):
-    """
-    """
     return order_classification_config_model.get_order_classification_config_by_id(
         session=session,
         config_id=config_id,
@@ -36,8 +43,6 @@ def get_classification_config_by_user_id(
     session: Session = Depends(get_session),
     current_user: UserRead = Depends(get_current_user),
 ):
-    """
-    """
     return order_classification_config_model.get_order_classification_config_by_user_id(
         session=session,
         user_id=user_id,
@@ -56,14 +61,17 @@ def classify_order(
     """
     Вебхук для обновления аварийности заявки в Домиленд.
     """
-    order_classification_response = order_classification_model.classify_order(
+
+    logger.debug(f"New Order Classification request body:\n{body}")
+
+    model_response = order_classification_model.classify_order(
         body=body,
         client=client,
     )
 
     response_status = status.HTTP_200_OK
-    if order_classification_response.is_error:
+    if model_response.is_error:
         response_status = status.HTTP_400_BAD_REQUEST
-
     response.status_code = response_status
-    return order_classification_response
+
+    return model_response
